+++ src/components/game/NextBallsPreview.tsx
@@ -1,5 +1,6 @@
 import React from "react";
-import type { BallColor } from "../../game/constants";
+import type { BallColor } from "../../game/types";
+import { getBallColor } from "../../utils/helpers";
 
 interface NextBallsPreviewProps {
   nextBalls: BallColor[];
@@ -11,7 +12,8 @@ const NextBallsPreview: React.FC<NextBallsPreviewProps> = ({ nextBalls }) => {
       {nextBalls.map((color, i) => (
         <span
           key={i}
-          className={`block w-ball h-ball rounded-full border-2 border-game-border-default text-center align-middle bg-ball-${color}`}
+          className={`block w-ball h-ball rounded-full border-2 border-game-border-default text-center align-middle`}
+          style={{ backgroundColor: getBallColor(color) }}
           title={color}
         />
       ))}
