+++ src/game/types/index.ts
@@ -1,5 +1,6 @@
-import { type BallColor as _BallColor } from "../config";
-export type BallColor = _BallColor;
+import { BALL_COLORS } from "../config";
+
+export type BallColor = typeof BALL_COLORS[number];
 
 export interface Ball {
   color: BallColor;
@@ -62,4 +63,51 @@ export interface GameActions {
   handleCellLeave: () => void;
   handleNewGameFromDialog: () => void;
   handleCloseDialog: () => void;
+}
+
+// Game Phase Management Types
+export interface GamePhase {
+  type: "idle" | "moving" | "popping" | "converting" | "gameOver";
+  data?: Record<string, unknown>;
+}
+
+export interface MoveResult {
+  newBoard: Cell[][];
+  linesFormed: boolean;
+  ballsRemoved?: [number, number][];
+  pointsEarned?: number;
+  nextBalls?: BallColor[];
+}
+
+export interface ConversionResult {
+  newBoard: Cell[][];
+  nextBalls: BallColor[];
+  gameOver: boolean;
+}
+
+// Board State Types
+export interface BoardState {
+  board: Cell[][];
+  nextBalls: BallColor[];
+  setBoard: (board: Cell[][]) => void;
+  setNextBalls: (nextBalls: BallColor[], board?: Cell[][]) => void;
+}
+
+// Animation Types
+export interface AnimationState {
+  movingBall: { color: BallColor; path: [number, number][] } | null;
+  movingStep: number;
+  poppingBalls: Set<string>;
+  setMovingBall: (ball: { color: BallColor; path: [number, number][] } | null) => void;
+  setMovingStep: (step: number) => void;
+  setPoppingBalls: (balls: Set<string>) => void;
+}
+
+// Timer Types
+export interface TimerState {
+  timer: number;
+  timerActive: boolean;
+  setTimerActive: (active: boolean) => void;
+  resetTimer: () => void;
+  onActivity: () => void;
 }
